Added .md for bug reporting. Fixed some minor inconsistencies. The project development may continue its journey! Yohoo!

+++ src/main/java/wordwizard/models/VocabularyStats.java
@@ -5,7 +5,7 @@ import java.util.Map;
 public record VocabularyStats(
         int totalWords,
         int totalDefinitions,
-        double avgDefinitionsPerWord,
+        long avgDefinitionsPerWord,
         Map<String, Integer> wordsByPos,
         Map<String, Integer> wordsByTheme,
         int totalThemes,
